+++ takezo-portfolio/app/projects/data/cyberProjects.js
@@ -1,0 +1,11 @@
+const cyberProjects = [
+	{
+		name: "WIP",
+		description: "Coming Soon",
+		tech: "???",
+		src: "/assets/cyber-bg.svg",
+	},
+];
+
+export default cyberProjects;
+
