+++ takezo-portfolio/app/projects/data/devSkills.js
@@ -1,0 +1,9 @@
+const devSkills = {
+	languages: "HTML, CSS, JavaScript, C, C++, C#, Python, Java",
+	databases: "MongoDB, PostgreSQL, MySQL",
+	frameworks: "ReactJs, AngularJs, Next.js, Tailwind, Bootstrap 5",
+	devTools: "Git, Github, Figma, Postman, VS Code",
+};
+
+export default devSkills;
+
