+++ takezo-portfolio/app/projects/page.jsx
@@ -10,125 +10,10 @@ import Link from "next/link";
 // context
 import { useTheme } from "../../context/ThemeContext";
 
-const cyberProjects = [
-	{
-		name: "WIP",
-		description: "Coming Soon",
-		tech: "???",
-		src: "/assets/cyber-bg.svg",
-	},
-];
-const devProjects = [
-	{
-		name: "BG FILM",
-		description:
-			"BG Film website is a site for our group project in school about multimedia. My role here is Producer and Webmaster, which is why I created this website. No frameworks were used because I hadn’t explored them at that time.",
-		languages: "HTML, CSS, JS",
-		src: "/assets/bgfilm.png",
-		link: "https://bgfilm.vercel.app/",
-	},
-	{
-		name: "PlatzDodge",
-		description:
-			"It's group project, horizontal endless dodging game inspired by raylib game examples Space Invaders and Dr. turtle & Mr. gamera.",
-		languages: "C, Raylib",
-		src: "/assets/platzdodge.png",
-		link: "https://github.com/Takezooo/Platz-Dodge",
-	},
-	{
-		name: "Decima",
-		description:
-			"This was a game dev school group project. The title comes from the word decima, meaning 'tenth,' as the game is based on the ten circles of hell. It’s a side-scrolling platformer made with free assets that I modified for color, along with free music.",
-		languages: "C#, Unity",
-		src: "/assets/decima.png",
-		link: "https://dtakezo.itch.io/decima",
-	},
-	{
-		name: "League of Legends Smart Overlay",
-		description: "Org Repo",
-		languages: "Riot API",
-		src: "/assets/dev-bg.svg",
-		link: "/projects",
-	},
-	{
-		name: "Valorant Manually-controlled Overlay",
-		description: "Org Repo",
-		languages: "React, Socket.io",
-		src: "/assets/dev-bg.svg",
-		link: "/projects",
-	},
-	{
-		name: "Mobile Legends Manually-controlled Overlay",
-		description: "Org Repo",
-		languages: "React, Socket.io",
-		src: "/assets/dev-bg.svg",
-		link: "/projects",
-	},
-	{
-		name: "Call of Duty: Mobile Manually-controlled Overlay   ",
-		description: "Org Repo",
-		languages: "React, Socket.io",
-		src: "/assets/dev-bg.svg",
-		link: "/projects",
-	},
-	{
-		name: "iZip",
-		description:
-			"Group Project, forked repo (ethanbrimhall/kahoot-clone-nodejs ), Kahoot clone",
-		languages: "JavaScript, Nodejs",
-		src: "/assets/dev-bg.svg",
-		link: "/projects",
-	},
-	{
-		name: "Bookstore System",
-		description: "Practice Project",
-		languages: "MERN Stack",
-		src: "/assets/dev-bg.svg",
-		link: "/projects",
-	},
-	{
-		name: "Library Management System",
-		description: "Group Project, unfinished",
-		languages: "Java",
-		src: "/assets/dev-bg.svg",
-		link: "/projects",
-	},
-	{
-		name: "ATM Project with Flashdrive",
-		description: "School Project",
-		languages: "C",
-		src: "/assets/dev-bg.svg",
-		link: "/projects",
-	},
-	{
-		name: "iRenta",
-		description: "Capstone, Airbnb clone?",
-		languages: "MERN Stack",
-		src: "/assets/dev-bg.svg",
-		link: "/projects",
-	},
-	{
-		name: "Takezo's Blog Site",
-		description: "WIP",
-		languages: "???",
-		src: "/assets/dev-bg.svg",
-		link: "/projects",
-	},
-	{
-		name: "Breeding Sim",
-		description: "WIP",
-		languages: "Unity, C#, PostgreSQL, Node",
-		src: "/assets/dev-bg.svg",
-		link: "/projects",
-	},
-];
-
-const devSkills = {
-	languages: "HTML, CSS, JavaScript, C, C++, C#, Python, Java",
-	databases: "MongoDB, PostgreSQL, MySQL",
-	frameworks: "ReactJs, AngularJs, Next.js, Tailwind, Bootstrap 5",
-	devTools: "Git, Github, Figma, Postman, VS Code",
-};
+// data
+import cyberProjects from "./data/cyberProjects";
+import devProjects from "./data/devProjects";
+import devSkills from "./data/devSkills";
 
 const Projects = () => {
 	const { theme } = useTheme();
